# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HarvestCore API - Commodity-backed tokenization and lending platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthenticated
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"

  # ── USERS / KYC ──────────────────────────────────────────
...
  /inventory/intakes:
    get:
      operationId: listIntakes
      tags: [inventory]
      summary: List all intake records
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, graded, weighed, verified, anchored, rejected]
        - name: warehouseId
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of intakes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeListResponse"
    post:
      operationId: createIntake
      tags: [inventory]
      summary: Create a new intake record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIntakeRequest"
      responses:
        "201":
          description: Intake created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeRecord"
  /inventory/intakes/{intakeId}:
    get:
      operationId: getIntake
      tags: [inventory]
      summary: Get intake record
      parameters:
        - name: intakeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Intake record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeRecord"
    patch:
      operationId: updateIntake
      tags: [inventory]
      summary: Update intake (grade/weight)
      parameters:
        - name: intakeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIntakeRequest"
      responses:
        "200":
          description: Updated intake
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeRecord"
  /inventory/intakes/{intakeId}/approve:
    post:
      operationId: approveIntake
      tags: [inventory]
      summary: Checker approves intake (Maker-Checker)
      parameters:
        - name: intakeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Intake approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeRecord"
  /inventory/intakes/{intakeId}/reject:
    post:
      operationId: rejectIntake
      tags: [inventory]
      summary: Checker rejects intake
      parameters:
        - name: intakeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
              required: [reason]
      responses:
        "200":
          description: Intake rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeRecord"
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]

    # ── USER ─────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [farmer, warehouse_op, checker, lender, admin]
        orgId:
          type: string
          nullable: true
        kycStatus:
          type: string
          enum: [pending, approved, rejected]
        phone:
          type: string
          nullable: true
        walletAddress:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, kycStatus, createdAt]
...
    IntakeRecord:
      type: object
      properties:
        id:
          type: string
        farmerId:
          type: string
        farmerName:
          type: string
        warehouseId:
          type: string
        warehouseName:
          type: string
        commodity:
          type: string
        variety:
          type: string
          nullable: true
        weightKg:
          type: number
        moisturePercent:
          type: number
        grade:
          type: string
          enum: [A, B, C, D]
          nullable: true
        grnNumber:
          type: string
          nullable: true
        ewrsRegistryId:
          type: string
          nullable: true
        iotaHash:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, graded, weighed, verified, anchored, rejected]
        checkerNotes:
          type: string
          nullable: true
        checkedBy:
          type: string
          nullable: true
        tokenId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, farmerId, farmerName, warehouseId, warehouseName, commodity, weightKg, moisturePercent, status, createdAt, updatedAt]
    IntakeListResponse:
      type: object
      properties:
        intakes:
          type: array
          items:
            $ref: "#/components/schemas/IntakeRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [intakes, total, page, limit]
    CreateIntakeRequest:
      type: object
      properties:
        farmerId:
          type: string
        warehouseId:
          type: string
        commodity:
          type: string
        variety:
          type: string
        weightKg:
          type: number
        moisturePercent:
          type: number
      required: [farmerId, warehouseId, commodity, weightKg, moisturePercent]
    UpdateIntakeRequest:
      type: object
      properties:
        grade:
          type: string
          enum: [A, B, C, D]
        weightKg:
          type: number
        moisturePercent:
          type: number
        grnNumber:
          type: string
        checkerNotes:
          type: string

    # ── TOKENS ────────────────────────────────────────────────
...