# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HarvestCore API - Commodity-backed tokenization and lending platform
 * OpenAPI spec version: 0.1.0
 */
import type { UserKycStatus } from "./userKycStatus";
import type { UserRole } from "./userRole";

export interface User {
  id: string;
  email: string;
  name: string;
  role: UserRole;
  orgId?: string | null;
  kycStatus: UserKycStatus;
  phone?: string | null;
  walletAddress?: string | null;
  createdAt: Date;
}